% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity-data-objects.R
\docType{data}
\name{ganges_params}
\alias{ganges_params}
\title{Ganges calibrated parameters}
\format{
An object of class \code{data.frame} with 4 rows and 2 columns.
}
\usage{
ganges_params
}
\description{
This data contains calibrated parameters for modeling Ganges salinity at Khulna
}
\keyword{datasets}
