% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity-data-objects.R
\docType{data}
\name{ganges_streamflow}
\alias{ganges_streamflow}
\title{Ganges streamflow data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 362 rows and 5 columns.
}
\usage{
ganges_streamflow
}
\description{
This data contains daily average streamflow data over the time period of the treaty (after 1998)
as well as synthetic data from 2006 without a treaty.
}
\keyword{datasets}
