% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity-SLR.R
\name{get_d_SLR}
\alias{get_d_SLR}
\title{Get parameter d for SLR}
\usage{
get_d_SLR(SLR_m, probs = 0.5)
}
\arguments{
\item{SLR_m}{vector of sea level rise in meters}

\item{probs}{Determines the quantile for d where 1 returns max(d), 0 returns min(d)}
}
\value{
Returns the associated values as log(d).
}
\description{
Get parameter d for a given SLR and quantil
}
\details{
Given the uncertainty of the parameter d under sea level rise, we specify this parameter
within a range from min(d) to max(d). The \code{probs} parameter determines which quantile
to extract within this range, with 0 being the lowest, 0.5 the median, and 1 the maximum.
This uniform distribution corresponds to the parameter d, not log(d) (although log(d) is returned).
Note that either SLR_m or probs can be a vector of length greater than 1, but not both.
}
\examples{
get_d_SLR(SLR_m = 0.5)
}
