% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity-SLR.R
\name{get_ganges_SLR_gw_ratios}
\alias{get_ganges_SLR_gw_ratios}
\title{Get groundwater volume ratios}
\usage{
get_ganges_SLR_gw_ratios(SLR_m)
}
\arguments{
\item{SLR_m}{vector of sea level rise in meters}
}
\description{
Get groundwater-connected volume ratios under sea level rise
}
\examples{
library(ggplot2)
gw_ratios <- get_ganges_SLR_gw_ratios(SLR_m = seq(0,1,by = 0.05))
ggplot(gw_ratios) +
  geom_line(aes(SLR_m, R_E, color = "R_E")) +
  geom_line(aes(SLR_m, R_V, color = "R_V")) + ylab("Ratio")
}
\keyword{internal}
