% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity-SLR.R
\name{get_SLR_channel_ratios}
\alias{get_SLR_channel_ratios}
\title{Get channel volume ratios}
\usage{
get_SLR_channel_ratios(SLR_m)
}
\arguments{
\item{SLR_m}{vector of sea level rise in meters}
}
\description{
Get channel volume ratios under sea level rise
}
\examples{
channel_ratios <- get_SLR_channel_ratios(SLR_m = seq(0,1,by = 0.05))
ggplot(channel_ratios) +
  geom_line(aes(SLR_m, R_E, color = "R_E")) +
  geom_line(aes(SLR_m, R_V, color = "R_V")) + ylab("Ratio")
}
