% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity-model-functions.R
\name{sim_salin_annual}
\alias{sim_salin_annual}
\title{Simulate salinity, re-initializing each year}
\usage{
sim_salin_annual(Q_df, v)
}
\arguments{
\item{Q_df}{A \code{data.frame} containing \code{Q_cumec} and \code{year} columns. The rows must be in order of consecutive days.}

\item{v}{Vector of length 4 containing log parameter values: \code{log(a), log(b), log(d), and log(C_d)}}
}
\description{
Simulate salinity, re-initializing each year
}
\examples{
library(ggplot2)

# Create new data.frame with streamflow input variables
streamflow_df <- ganges_streamflow

# Simulate salinity each year
streamflow_df$S_ppm <- sim_salin_annual(ganges_streamflow, ganges_params$param)

# Plot the output
ggplot(streamflow_df) + geom_line(aes(yday,S_ppm, color = group))
}
