% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity-prep.R
\name{sim_salin_prep}
\alias{sim_salin_prep}
\title{Join partial timeseries of synthetic streamflow with observed streamflow}
\usage{
sim_salin_prep(Q_ts = NULL, pred_dates = NULL, Q_obs_df = NULL)
}
\arguments{
\item{Q_ts}{Vector of streamflow values}

\item{pred_dates}{vector of dates for Q_ts}

\item{Q_obs_df}{Data.frame that contains columns with date and observed Q_cumec}
}
\description{
Prepare synthetic salinity timeseries by combining observed streamflow
with synthetic streamflow from only part of the year
}
\keyword{internal}
