% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_salinity_model.R
\name{sse_sim_salin}
\alias{sse_sim_salin}
\title{Optmize parameters}
\usage{
sse_sim_salin(v, hydro_data, sse = TRUE, salin_init = NULL, salin_min = 100)
}
\arguments{
\item{v}{logged parameter as a vector: log(a), log(b), log(d), log(C_d)}

\item{hydro_data}{data.frame containing \code{S_ppm} and \code{Q_cumec} columns for single year}

\item{sse}{logical: if \code{TRUE} return SSE, otherwise return vector of salinity}

\item{salin_init}{initial salinity. If set to NULL, function will preferentially take 1st salinity observation
in hydro_data or salin_min}

\item{salin_min}{minimum salinity, ie river salinity inflow to control volume}
}
\description{
Optmize parameters
}
\details{
This function returns the sum of squared error for \code{sim_salin()} using \code{hydro_data$S_ppm}
as a benchmark. Note that any elements of \code{hydro_data$S_ppm} that are \code{NA} will be ignored in
the calibration, meaning that it's possible to calibrate on a subset of dates by setting observed
salinity (\code{hydro_data$S_ppm}) to \code{NA} except on those dates where the model should be calibrated.
}
\examples{
\dontrun{
hydro_data <- ganges_streamflow[ganges_streamflow$date < "2000-01-01",]
# Output salinity in ppm
hydro_data$S_ppm_model <- sim_salin(hydro_data$Q_cumec, ganges_params$param)
hydro_data$S_ppm <- hydro_data$S_ppm_model * runif(nrow(hydro_data), min = 0.9, max = 1.1)
v <- ganges_params$param
sse_sim_salin(v, hydro_data, sse = TRUE)
}
}
\keyword{internal}
