% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity-calibrate-model.R
\name{sse_wrapper}
\alias{sse_wrapper}
\title{Wrapper around sse}
\usage{
sse_wrapper(v_calibrate, v, hydro_data, sse = TRUE)
}
\arguments{
\item{v_calibrate}{These values replace NA values in v}

\item{v}{Model parameters as log(a), lob(b), log(d), log(C_d). Set to NA those which
should be replaced by \code{v_calibrate}}

\item{hydro_data}{data.frame containing \code{S_ppm}, \code{Q_cumec}, and \code{year} columns}

\item{sse}{logical, if TRUE will return SSE. Otherwise, returns full data.frame of results}
}
\description{
Wrapper around sse
}
\details{
This functions calculates SSE (sum of squared error), comparing modeled salinity output to the
calibration data (\code{hydro_data$S_ppm}). Salinity is modeled using daily streamflow
(\code{hydro_data$Q_cumec}) and the parameters \code{v} -- and those values of \code{v} that are
NA are replaced by values in \code{v_calibrate}.
}
\examples{
hydro_data_orig <- ganges_streamflow
v <- ganges_params$param
# Output salinity in ppm
hydro_data <- sim_salin_annual(hydro_data_orig, v)

# add random error to salinity output
set.seed(100)
hydro_data$S_ppm_orig <- hydro_data$S_ppm
hydro_data$S_ppm <- hydro_data$S_ppm_orig * runif(nrow(hydro_data), min = 0.9, max = 1.1)
hydro_data$year <- as.numeric(strftime(hydro_data$date,"\%Y"))

sse_all <- sse_wrapper(NULL, v, hydro_data)
sse_all

hydro_data$Cvect <- sse_wrapper(NULL, v, hydro_data, FALSE)$Cvect
sse_all_manual <- sum((hydro_data$S_ppm - hydro_data$Cvect)^2)
sse_all_manual

# Testing specific values
sse_wrapper(NULL, c(-12,v[2:4]), hydro_data)
}
