% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity-model-functions.R
\name{sim_salin_annual}
\alias{sim_salin_annual}
\title{Simulate salinity, re-initializing each year}
\usage{
sim_salin_annual(Q_df, v)
}
\arguments{
\item{Q_df}{A \code{data.frame} containing \code{Q_cumec} and \code{year} columns}

\item{v}{Vector of length 4 containing log parameter values: \code{log(a), log(b), log(d), and log(C_d)}}
}
\description{
Simulate salinity, re-initializing each year
}
\examples{
library(dplyr)
library(ggplot2)
salinity_results <- sim_salin_annual(ganges_streamflow, ganges_params$param)
salinity_results_joined <- ganges_streamflow \%>\%
  left_join(salinity_results[,c("date","S_ppm")], by = "date")
ggplot(salinity_results_joined) + geom_line(aes(yday,S_ppm, color = group))
}
